+++ src/main/java/cn/jf180/finance/dao/fin/dao/CreditOrderGoodsMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CreditOrderGoodsMapper {
-    int countByExample(CreditOrderGoodsExample example);
+    long countByExample(CreditOrderGoodsExample example);
 
     int deleteByExample(CreditOrderGoodsExample example);
